refactor(lab-runner): extract workspace listing/snapshots from sync god file
workspace/sync.rs was a 2163-line god file (audit god_file finding).
Extract the read-only listing/snapshot cluster — listing lab workspaces
(local + SSH), building the snapshot view (per-workspace metadata entries
with applied filters), and the workspace-scan command — into a focused
snapshots.rs submodule, separating inspection from the
materialization/prune/metadata write paths.

Convert the file to a directory module (git mv preserves history; the
module path workspace::sync is unchanged, so mod.rs paths stay). Clean
seam: only workspace_snapshots + list_workspaces are called externally
(both pub, re-exported), the cluster's sole outbound helper is
workspace_repo_from_path, and it shares shell_arg / RunnerKind / load /
WORKSPACE_METADATA_FILE (now pub(crate)). External pub contract
(workspace::{list_workspaces, workspace_snapshots}) unchanged.

Behavior-preserving: sync/mod.rs 2163 -> 1788 lines, snapshots.rs 402.
All 7 sync + 13 workspace snapshot tests pass.
